requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
pydantic==2.5.0
pydantic-settings==2.1.0
stripe==7.4.0
python-dotenv==1.0.0
email-validator==2.0.0
# Test dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2